Extract run-related args/actions
