+++ xtask/src/main.rs
@@ -5,7 +5,7 @@ use std::{
     time::Instant,
 };
 
-use anyhow::{Context as _, Result, bail, ensure};
+use anyhow::{Result, bail};
 use clap::{Args, Parser};
 use esp_metadata::{Chip, Config};
 use strum::IntoEnumIterator;
@@ -14,7 +14,6 @@ use xtask::{
     Version,
     cargo::{CargoAction, CargoArgsBuilder},
     commands::*,
-    firmware::Metadata,
 };
 
 // ----------------------------------------------------------------------------
@@ -25,9 +24,14 @@ enum Cli {
     /// Build-related subcommands
     #[clap(subcommand)]
     Build(Build),
+    /// Run-related subcommands
+    #[clap(subcommand)]
+    Run(Run),
 
     /// Bump the version of the specified package(s).
     BumpVersion(BumpVersionArgs),
+    /// Perform (parts of) the checks done in CI
+    Ci(CiArgs),
     /// Format all packages in the workspace with rustfmt
     #[clap(alias = "format-packages")]
     FmtPackages(FmtPackagesArgs),
@@ -35,47 +39,8 @@ enum Cli {
     LintPackages(LintPackagesArgs),
     /// Attempt to publish the specified package.
     Publish(PublishArgs),
-    /// Run doctests for specified chip and package.
-    #[clap(alias = "run-doc-test")]
-    RunDocTests(ExampleArgs),
-    /// Run the given example for the specified chip.
-    RunExample(BuildExamplesArgs),
-    /// Run all applicable tests or the specified test for a specified chip.
-    RunTests(BuildTestsArgs),
-    /// Run all ELFs in a folder.
-    RunElfs(RunElfArgs),
-    /// Perform (parts of) the checks done in CI
-    Ci(CiArgs),
     /// Generate git tags for all new package releases.
     TagReleases(TagReleasesArgs),
-}
-
-#[derive(Debug, Args)]
-struct ExampleArgs {
-    /// Package whose examples we which to act on.
-    #[arg(value_enum)]
-    package: Package,
-    /// Chip to target.
-    #[arg(value_enum)]
-    chip: Chip,
-    /// Optional example to act on (all examples used if omitted)
-    example: Option<String>,
-    /// Build examples in debug mode only
-    #[arg(long)]
-    debug: bool,
-}
-
-#[derive(Debug, Args)]
-struct TestArgs {
-    /// Chip to target.
-    #[arg(value_enum)]
-    chip: Chip,
-    /// Optional test to act on (all tests used if omitted)
-    #[arg(short = 't', long)]
-    test: Option<String>,
-    /// Repeat the tests for a specific number of times.
-    #[arg(long)]
-    repeat: Option<usize>,
 }
 
 #[derive(Debug, Args)]
@@ -89,6 +54,13 @@ struct BumpVersionArgs {
 }
 
 #[derive(Debug, Args)]
+struct CiArgs {
+    /// Chip to target.
+    #[arg(value_enum)]
+    chip: Chip,
+}
+
+#[derive(Debug, Args)]
 struct FmtPackagesArgs {
     /// Run in 'check' mode; exists with 0 if formatted correctly, 1 otherwise
     #[arg(long)]
@@ -97,15 +69,6 @@ struct FmtPackagesArgs {
     /// Package(s) to target.
     #[arg(value_enum, default_values_t = Package::iter())]
     packages: Vec<Package>,
-}
-
-#[derive(Debug, Args)]
-struct GenerateEfuseFieldsArgs {
-    /// Path to the local ESP-IDF repository.
-    idf_path: PathBuf,
-    /// Chip to build eFuse fields table for.
-    #[arg(value_enum)]
-    chip: Chip,
 }
 
 #[derive(Debug, Args)]
@@ -132,22 +95,6 @@ struct PublishArgs {
     /// Do not pass the `--dry-run` argument, actually try to publish.
     #[arg(long)]
     no_dry_run: bool,
-}
-
-#[derive(Debug, Args)]
-struct RunElfArgs {
-    /// Which chip to run the tests for.
-    #[arg(value_enum)]
-    chip: Chip,
-    /// Path to the ELFs.
-    path: PathBuf,
-}
-
-#[derive(Debug, Args)]
-struct CiArgs {
-    /// Chip to target.
-    #[arg(value_enum)]
-    chip: Chip,
 }
 
 #[derive(Debug, Args)]
@@ -187,236 +134,25 @@ fn main() -> Result<()> {
             ),
         },
 
+        // Run-related subcommands:
+        Cli::Run(run) => match run {
+            Run::DocTests(args) => run_doc_tests(&workspace, args),
+            Run::Elfs(args) => run_elfs(args),
+            Run::Example(args) => examples(&workspace, args, CargoAction::Run),
+            Run::Tests(args) => tests(&workspace, args, CargoAction::Run),
+        },
+
         Cli::BumpVersion(args) => bump_version(&workspace, args),
+        Cli::Ci(args) => run_ci_checks(&workspace, args),
         Cli::FmtPackages(args) => fmt_packages(&workspace, args),
         Cli::LintPackages(args) => lint_packages(&workspace, args),
         Cli::Publish(args) => publish(&workspace, args),
-        Cli::RunDocTests(args) => run_doc_tests(&workspace, args),
-        Cli::RunElfs(args) => run_elfs(args),
-        Cli::RunExample(args) => examples(&workspace, args, CargoAction::Run),
-        Cli::RunTests(args) => tests(&workspace, args, CargoAction::Run),
-        Cli::Ci(args) => run_ci_checks(&workspace, args),
         Cli::TagReleases(args) => tag_releases(&workspace, args),
     }
 }
 
 // ----------------------------------------------------------------------------
 // Subcommands
-
-pub fn examples(workspace: &Path, mut args: BuildExamplesArgs, action: CargoAction) -> Result<()> {
-    // Ensure that the package/chip combination provided are valid:
-    args.package.validate_package_chip(&args.chip)?;
-
-    // If the 'esp-hal' package is specified, what we *really* want is the
-    // 'examples' package instead:
-    if args.package == Package::EspHal {
-        log::warn!(
-            "Package '{}' specified, using '{}' instead",
-            Package::EspHal,
-            Package::Examples
-        );
-        args.package = Package::Examples;
-    }
-
-    // Absolute path of the package's root:
-    let package_path = xtask::windows_safe_path(&workspace.join(args.package.to_string()));
-
-    let example_path = match args.package {
-        Package::Examples | Package::QaTest => package_path.join("src").join("bin"),
-        Package::HilTest => package_path.join("tests"),
-        _ => package_path.join("examples"),
-    };
-
-    // Load all examples which support the specified chip and parse their metadata:
-    let mut examples = xtask::firmware::load(&example_path)?
-        .iter()
-        .filter_map(|example| {
-            if example.supports_chip(args.chip) {
-                Some(example.clone())
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-
-    // Sort all examples by name:
-    examples.sort_by_key(|a| a.binary_name());
-
-    // Execute the specified action:
-    match action {
-        CargoAction::Build(out_path) => build_examples(args, examples, &package_path, &out_path),
-        CargoAction::Run if args.example.is_some() => run_example(args, examples, &package_path),
-        CargoAction::Run => run_examples(args, examples, &package_path),
-    }
-}
-
-pub fn tests(workspace: &Path, args: BuildTestsArgs, action: CargoAction) -> Result<()> {
-    // Absolute path of the 'hil-test' package's root:
-    let package_path = xtask::windows_safe_path(&workspace.join("hil-test"));
-
-    // Determine the appropriate build target for the given package and chip:
-    let target = Package::HilTest.target_triple(&args.chip)?;
-
-    // Load all tests which support the specified chip and parse their metadata:
-    let mut tests = xtask::firmware::load(&package_path.join("tests"))?
-        .into_iter()
-        .filter(|example| example.supports_chip(args.chip))
-        .collect::<Vec<_>>();
-
-    // Sort all tests by name:
-    tests.sort_by_key(|a| a.binary_name());
-
-    // Execute the specified action:
-    if tests.iter().find(|test| test.matches(&args.test)).is_some() {
-        for test in tests.iter().filter(|test| test.matches(&args.test)) {
-            xtask::execute_app(
-                &package_path,
-                args.chip,
-                target,
-                test,
-                action.clone(),
-                args.repeat.unwrap_or(1),
-                false,
-            )?;
-        }
-        Ok(())
-    } else if args.test.is_some() {
-        bail!("Test not found or unsupported for the given chip")
-    } else {
-        let mut failed = Vec::new();
-        for test in tests {
-            if xtask::execute_app(
-                &package_path,
-                args.chip,
-                target,
-                &test,
-                action.clone(),
-                args.repeat.unwrap_or(1),
-                false,
-            )
-            .is_err()
-            {
-                failed.push(test.name_with_configuration());
-            }
-        }
-
-        if !failed.is_empty() {
-            bail!("Failed tests: {:#?}", failed);
-        }
-
-        Ok(())
-    }
-}
-
-fn run_example(
-    args: BuildExamplesArgs,
-    examples: Vec<Metadata>,
-    package_path: &Path,
-) -> Result<()> {
-    // Determine the appropriate build target for the given package and chip:
-    let target = args.package.target_triple(&args.chip)?;
-
-    // Filter the examples down to only the binary we're interested in, assuming it
-    // actually supports the specified chip:
-    let mut found_one = false;
-    for example in examples.iter().filter(|ex| ex.matches(&args.example)) {
-        found_one = true;
-        xtask::execute_app(
-            package_path,
-            args.chip,
-            target,
-            example,
-            CargoAction::Run,
-            1,
-            args.debug,
-        )?;
-    }
-
-    ensure!(
-        found_one,
-        "Example not found or unsupported for {}",
-        args.chip
-    );
-
-    Ok(())
-}
-
-fn run_examples(
-    args: BuildExamplesArgs,
-    examples: Vec<Metadata>,
-    package_path: &Path,
-) -> Result<()> {
-    // Determine the appropriate build target for the given package and chip:
-    let target = args.package.target_triple(&args.chip)?;
-
-    // Filter the examples down to only the binaries we're interested in
-    let mut examples: Vec<Metadata> = examples
-        .iter()
-        .filter(|ex| ex.supports_chip(args.chip))
-        .cloned()
-        .collect();
-    examples.sort_by_key(|ex| ex.tag());
-
-    let console = console::Term::stdout();
-
-    for example in examples {
-        let mut skip = false;
-
-        log::info!("Running example '{}'", example.output_file_name());
-        if let Some(description) = example.description() {
-            log::info!(
-                "\n\n{}\n\nPress ENTER to run example, `s` to skip",
-                description.trim()
-            );
-        } else {
-            log::info!("\n\nPress ENTER to run example, `s` to skip");
-        }
-
-        loop {
-            let key = console.read_key();
-
-            match key {
-                Ok(console::Key::Enter) => break,
-                Ok(console::Key::Char('s')) => {
-                    skip = true;
-                    break;
-                }
-                _ => (),
-            }
-        }
-
-        if !skip {
-            while !skip
-                && xtask::execute_app(
-                    package_path,
-                    args.chip,
-                    target,
-                    &example,
-                    CargoAction::Run,
-                    1,
-                    args.debug,
-                )
-                .is_err()
-            {
-                log::info!("Failed to run example. Retry or skip? (r/s)");
-                loop {
-                    let key = console.read_key();
-
-                    match key {
-                        Ok(console::Key::Char('r')) => break,
-                        Ok(console::Key::Char('s')) => {
-                            skip = true;
-                            break;
-                        }
-                        _ => (),
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
 
 fn bump_version(workspace: &Path, args: BumpVersionArgs) -> Result<()> {
     // Bump the version by the specified amount for each given package:
@@ -597,83 +333,6 @@ fn publish(workspace: &Path, args: PublishArgs) -> Result<()> {
     Ok(())
 }
 
-fn run_elfs(args: RunElfArgs) -> Result<()> {
-    let mut failed: Vec<String> = Vec::new();
-    for elf in fs::read_dir(&args.path)? {
-        let entry = elf?;
-
-        let elf_path = entry.path();
-        let elf_name = elf_path
-            .with_extension("")
-            .file_name()
-            .unwrap()
-            .to_string_lossy()
-            .to_string();
-
-        log::info!("Running test '{}' for '{}'", elf_name, args.chip);
-
-        let mut command = Command::new("probe-rs");
-        command.arg("run").arg(elf_path);
-
-        if args.chip == Chip::Esp32c2 {
-            command.arg("--speed").arg("15000");
-        };
-
-        command.arg("--verify");
-
-        let mut command = command.spawn().context("Failed to execute probe-rs")?;
-        let status = command
-            .wait()
-            .context("Error while waiting for probe-rs to exit")?;
-
-        log::info!("'{elf_name}' done");
-
-        if !status.success() {
-            failed.push(elf_name);
-        }
-    }
-
-    if !failed.is_empty() {
-        bail!("Failed tests: {:?}", failed);
-    }
-
-    Ok(())
-}
-
-fn run_doc_tests(workspace: &Path, args: ExampleArgs) -> Result<()> {
-    let chip = args.chip;
-
-    let package_name = args.package.to_string();
-    let package_path = xtask::windows_safe_path(&workspace.join(&package_name));
-
-    // Determine the appropriate build target, and cargo features for the given
-    // package and chip:
-    let target = args.package.target_triple(&chip)?;
-    let features = vec![chip.to_string(), "unstable".to_string()];
-
-    // We need `nightly` for building the doc tests, unfortunately:
-    let toolchain = if chip.is_xtensa() { "esp" } else { "nightly" };
-
-    // Build up an array of command-line arguments to pass to `cargo`:
-    let builder = CargoArgsBuilder::default()
-        .toolchain(toolchain)
-        .subcommand("test")
-        .arg("--doc")
-        .arg("-Zdoctest-xcompile")
-        .arg("-Zbuild-std=core,panic_abort")
-        .target(target)
-        .features(&features)
-        .arg("--release");
-
-    let args = builder.build();
-    log::debug!("{args:#?}");
-
-    // Execute `cargo doc` from the package root:
-    xtask::cargo::run(&args, &package_path)?;
-
-    Ok(())
-}
-
 fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
     let mut failure = false;
     let started_at = Instant::now();
@@ -695,7 +354,7 @@ fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
     // Check doc-tests
     run_doc_tests(
         workspace,
-        ExampleArgs {
+        ExamplesArgs {
             package: Package::EspHal,
             chip: args.chip,
             example: None,
@@ -727,7 +386,7 @@ fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
         // expects it
         examples(
             workspace,
-            BuildExamplesArgs {
+            ExamplesArgs {
                 package: Package::EspLpHal,
                 chip: args.chip,
                 example: None,
@@ -790,7 +449,7 @@ fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
     // Build (examples)
     examples(
         workspace,
-        BuildExamplesArgs {
+        ExamplesArgs {
             package: Package::Examples,
             chip: args.chip,
             example: None,
@@ -804,7 +463,7 @@ fn run_ci_checks(workspace: &Path, args: CiArgs) -> Result<()> {
     // Build (qa-test)
     examples(
         workspace,
-        BuildExamplesArgs {
+        ExamplesArgs {
             package: Package::QaTest,
             chip: args.chip,
             example: None,
